Enhanced light system -- darken tiles depending on distance from player
* Generate light level textures on the fly and add them to global texture atlas
* Player luminosity -- distance at which light obstruction is ignored, wip
* Skill exp is now stored as double, but is still displayed as an integer
* Resource skills character EXP is now multiplied by RES_EXP_COEF
* Hide mouse tooltip & map cursor when taking a screenshot + include milliseconds in screenshot name
* Move SpriteBatch to GameWorldRenderer
* Direction baseOrdinal() (idx of a 4-way direction, ordered counter-clockwise starting from RIGHT) optimization
* Doors are now craftable & placeable
* Wooden planks can now be broken with an axe
* Placeable tiles are now placed underneath the player instead of in front of him

** Crop bug fix WIP

+++ core/src/hitonoriol/madsand/entities/skill/SkillContainer.java
@@ -64,7 +64,7 @@ public class SkillContainer extends HashMap<Skill, SkillValue> {
 		return increaseSkill(skill, 1);
 	}
 
-	public boolean increaseSkill(Skill skill, int amt) {
+	public boolean increaseSkill(Skill skill, double amt) {
 		if (skill == Skill.None)
 			return false;
 
@@ -85,7 +85,7 @@ public class SkillContainer extends HashMap<Skill, SkillValue> {
 	}
 
 	public int getExp(Skill skill) {
-		return get(skill).exp;
+		return (int) get(skill).exp;
 	}
 
 	public int getExp() {
